agricultural_orchestrator/pyproject.toml
[project]
name = "agricultural-orchestrator"
version = "1.0.0"
description = "MVP Agricultural Intelligence Orchestrator Agent for Indian Farmers"
authors = [
    {name = "Agricultural AI Team", email = "team@example.com"}
]
dependencies = [
    "a2a-python>=0.1.0",
    "google-adk>=0.1.0",
    "google-generativeai>=0.3.0",
    "httpx>=0.25.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "asyncio-mqtt>=0.16.0",
    "nest-asyncio>=1.5.0",
    "pymongo"
]
requires-python = ">=3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0"
]
